refactor(frontend): split main.js into modules; dedupe markup; loop-drive presets
