+++ src/app/domain/web/templates/pages/explore.html.j2
@@ -107,39 +107,30 @@ SPDX-License-Identifier: Apache-2.0
 
     {# Vector storage calculator — pure client-side Vite JS #}
     <section id="panel-vector-calculator" data-ui-panel="vector-calculator" class="ui-panel min-w-0 p-4 lg:col-span-2">
-      <div class="flex flex-wrap items-start justify-between gap-3">
-        <div>
-          <h2 class="mb-1 text-sm font-semibold text-strong">Vector storage calculator</h2>
-          <p class="text-xs text-muted">Estimate Oracle 26ai vector payload, HNSW pool, and storage savings.</p>
-        </div>
-
+      <div>
+        <h2 class="mb-1 text-sm font-semibold text-strong">Vector storage calculator</h2>
+        <p class="text-xs text-muted">Estimate Oracle 26ai vector payload, HNSW pool, and storage savings.</p>
       </div>
 
       <div data-vector-calculator class="mt-4 grid gap-4 lg:grid-cols-2">
         <div class="space-y-4">
+          {% set vector_presets = [
+            {"label": "Gemini 768", "dimensions": 768, "format": "FLOAT32"},
+            {"label": "Gemini 3072", "dimensions": 3072, "format": "FLOAT32"},
+            {"label": "OpenAI 1536", "dimensions": 1536, "format": "FLOAT32"},
+            {"label": "Cohere 1024", "dimensions": 1024, "format": "FLOAT32"},
+            {"label": "Cohere 4096", "dimensions": 4096, "format": "FLOAT32"},
+          ] %}
           <div class="rounded-lg border border-border bg-surface-strong/55 p-3">
             <p class="mb-2 text-xs font-semibold uppercase text-muted">Model presets</p>
             <div class="grid gap-2 sm:grid-cols-3">
-              <button type="button" data-vector-preset data-preset-dimensions="768" data-preset-format="FLOAT32"
-                class="rounded-lg border border-border bg-surface px-3 py-2 text-left text-xs font-semibold text-base transition-colors hover:border-accent/40 hover:bg-accent-soft">
-                Gemini 768
-              </button>
-              <button type="button" data-vector-preset data-preset-dimensions="3072" data-preset-format="FLOAT32"
-                class="rounded-lg border border-border bg-surface px-3 py-2 text-left text-xs font-semibold text-base transition-colors hover:border-accent/40 hover:bg-accent-soft">
-                Gemini 3072
-              </button>
-              <button type="button" data-vector-preset data-preset-dimensions="1536" data-preset-format="FLOAT32"
-                class="rounded-lg border border-border bg-surface px-3 py-2 text-left text-xs font-semibold text-base transition-colors hover:border-accent/40 hover:bg-accent-soft">
-                OpenAI 1536
-              </button>
-              <button type="button" data-vector-preset data-preset-dimensions="1024" data-preset-format="FLOAT32"
-                class="rounded-lg border border-border bg-surface px-3 py-2 text-left text-xs font-semibold text-base transition-colors hover:border-accent/40 hover:bg-accent-soft">
-                Cohere 1024
-              </button>
-              <button type="button" data-vector-preset data-preset-dimensions="4096" data-preset-format="FLOAT32"
-                class="rounded-lg border border-border bg-surface px-3 py-2 text-left text-xs font-semibold text-base transition-colors hover:border-accent/40 hover:bg-accent-soft">
-                Cohere 4096
-              </button>
+              {% for p in vector_presets %}
+                <button type="button" data-vector-preset data-preset-dimensions="{{ p.dimensions }}"
+                  data-preset-format="{{ p.format }}"
+                  class="rounded-lg border border-border bg-surface px-3 py-2 text-left text-xs font-semibold transition-colors hover:border-accent/40 hover:bg-accent-soft">
+                  {{ p.label }}
+                </button>
+              {% endfor %}
             </div>
           </div>
