+++ src/app/domain/web/templates/partials/_flash.html.j2
@@ -3,10 +3,11 @@
  SPDX-License-Identifier: Apache-2.0
 #}
 
+{% set flash_colors = {"success": "success", "error": "danger", "warning": "accent-strong"} %}
 <div id="flash-region" hx-swap-oob="afterbegin" class="mx-auto max-w-5xl px-4 pt-3 space-y-2">
   {% for f in get_flashes() %}
     <div role="alert" class="flash flash-{{ f.category }} rounded-lg border border-border bg-surface px-3 py-2 text-sm">
-      <span class="font-medium text-{{ 'success' if f.category == 'success' else ('danger' if f.category == 'error' else ('accent-strong' if f.category == 'warning' else 'muted')) }}">
+      <span class="font-medium text-{{ flash_colors.get(f.category, "muted") }}">
         {{ f.message }}
       </span>
     </div>
